feat(cohorts): run the summary MV backfill as an idempotent migration
The backfill was a hand-run script: it appended to the MVs, so it needed
an exact --until to avoid double counting rows the live trigger had
already written, and it refused to run without one. It also mixed
topologies, resolving its target to the node-local <mv>_replicated while
reading from the Distributed events.

Make it migration 21 and rebuild rather than append. Each month's
partition is dropped and rebuilt from events, which is idempotent by
construction: an evicted migration pod is resumed by running again, and a
month that was double counted heals on the next pass. That in turn removes
the need to know when the MVs were created, so no operator input is
required and a small deployment gets working cohorts with no flags.

The bound is now read per month, straight after that month's drop. A run
can take hours and the current month is dropped at the end of it, so a
single bound taken at the start would delete every row the trigger wrote
during the run and then decline to rebuild them. Boundary strings keep
milliseconds for the same reason: created_at is DateTime64(3), and
truncating the final batch to whole seconds drops the events in the
boundary second after their trigger rows are gone.

Reads and writes both go through the Distributed tables, so one run from
one node covers every shard. DROP PARTITION is the exception, since the
Distributed engine rejects partitioning, so it goes ON CLUSTER against the
local table.

Rebuilding a large events table unattended is still not wanted, so it
steps aside above COHORT_BACKFILL_MAX_EVENTS rows and prints the manual
command. The same file remains directly executable for supervised runs.

Verified on standalone 26.1.3 and a keeper-backed 2-shard cluster 25.3:
fresh install no-ops, history plus live-trigger overlap matches ground
truth computed from events for both tables, one run writes both shards,
re-runs from either node are stable, and a deliberately double-counted
month returns to exact.
